+++ packages/react/fixtures/footerProps.js
@@ -1,0 +1,83 @@
+import { image } from '../../../utils';
+
+export default {
+  columns: [
+    {
+      title: 'Wings',
+      rows: [
+        {
+          type: 'image',
+          src: image(320, 100),
+          alt: 'Wings',
+          url: 'https://wings.dev',
+        },
+      ],
+    },
+    {
+      title: 'Contact',
+      rows: [
+        {
+          title: 'Address',
+          type: 'text',
+          content: `
+          67 Blahblah
+          Some City
+          Somewhere
+          `,
+        },
+        {
+          title: 'Email',
+          type: 'link',
+          url: 'mailto:info@wings.dev',
+          content: 'info@wings.dev',
+        },
+      ],
+    },
+    {
+      title: 'Social',
+      rows: [
+        {
+          type: 'social',
+          profiles: [
+            {
+              platform: 'twitter',
+              url: 'https://twitter.com/wingscms',
+            },
+            {
+              platform: 'facebook',
+              url: 'https://example.com',
+            },
+            {
+              platform: 'instagram',
+              url: 'https://example.com',
+            },
+            {
+              platform: 'linkedin',
+              url: 'https://example.com',
+            },
+          ],
+        },
+      ],
+    },
+    {
+      title: 'Text',
+      rows: [
+        {
+          type: 'text',
+          content:
+            'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt.',
+        },
+      ],
+    },
+    {
+      title: 'Button',
+      rows: [
+        {
+          type: 'button',
+          content: 'A button',
+          url: 'https://example.com',
+        },
+      ],
+    },
+  ],
+};