+++ packages/react/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.5.0](https://github.com/wingscms/wings/compare/@wingscms/react@1.4.0...@wingscms/react@1.5.0) (2020-08-07)
+
+
+### Bug Fixes
+
+* **react:** use textColor in nodes card ([#533](https://github.com/wingscms/wings/issues/533)) ([ffeb802](https://github.com/wingscms/wings/commit/ffeb80267bcb39ad7ade702775a77d7d8e73f8f5))
+
+
+### Features
+
+* **react:** add option to hide Entry header ([#534](https://github.com/wingscms/wings/issues/534)) ([be7a573](https://github.com/wingscms/wings/commit/be7a57389695f0e9f2460a9798f25de1d61de4bc))
+
+
+
+
+
+# [1.4.0](https://github.com/wingscms/wings/compare/@wingscms/react@1.3.1...@wingscms/react@1.4.0) (2020-08-04)
+
+
+### Features
+
+* button and accordion theming improvements ([#522](https://github.com/wingscms/wings/issues/522)) ([cf3c883](https://github.com/wingscms/wings/commit/cf3c883988ed42d786a6da559d0a142520d09aa9))
+* **components/react:** Portal; Overlay; Dialog; Language Switcher; Menu ([#529](https://github.com/wingscms/wings/issues/529)) ([6ac764f](https://github.com/wingscms/wings/commit/6ac764ff1afbffd90440399c3947c5e93b47b941))
+
+
+
+
+
+## [1.3.1](https://github.com/wingscms/wings/compare/@wingscms/react@1.3.0...@wingscms/react@1.3.1) (2020-07-28)
+
+
+### Bug Fixes
+
+* **react:** campaign heading sizes ([#525](https://github.com/wingscms/wings/issues/525)) ([3d00f7e](https://github.com/wingscms/wings/commit/3d00f7ea16ad72a78a042fd55e7a2d0c52ce0430))
+* **react:** Use Text/Heading components for campaign form texts ([#520](https://github.com/wingscms/wings/issues/520)) ([b2361a6](https://github.com/wingscms/wings/commit/b2361a6d4a3e437c7d9dbce9a305e0ff61b89842))
+
+
+
+
+
 # [1.3.0](https://github.com/wingscms/wings/compare/@wingscms/react@1.2.3...@wingscms/react@1.3.0) (2020-07-14)